Move gulpfile up
